feat: adds service factory
