+++ .pre-commit-config.yaml
@@ -25,7 +25,7 @@ repos:
     rev: v0.19.1
     hooks:
       - id: slotscheck
-        exclude: test_*|docs|migrations|scripts
+        exclude: test_*|docs|src/app/db/migrations|tools|manage.py
         entry: env PYTHONPATH=src slotscheck
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: v1.14.1
